refactor: extract SpinningBox component
